PHASE_1/API_SourceCode/requirements.txt
astroid==2.2.3
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-cors-headers==2.5.1
django-filter==2.0.0
djangorestframework==3.9.2
djangorestframework-filters==0.10.2.post0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
eventlet==0.24.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.10
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
openapi-codec==1.3.2
psycopg2-binary==2.7.7
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
rope==0.12.0
ruamel.yaml==0.15.89
simplejson==3.16.0
six==1.12.0
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
